rename token package to alter
