+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/asserts/segment/table/AlterTableAssert.java
@@ -22,8 +22,8 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.apache.shardingsphere.sql.parser.integrate.asserts.SQLCaseAssertContext;
 import org.apache.shardingsphere.sql.parser.integrate.jaxb.segment.impl.table.ExpectedAlterTable;
-import org.apache.shardingsphere.sql.parser.integrate.jaxb.segment.impl.token.ExpectedColumnDefinition;
-import org.apache.shardingsphere.sql.parser.integrate.jaxb.segment.impl.token.ExpectedColumnPosition;
+import org.apache.shardingsphere.sql.parser.integrate.jaxb.segment.impl.alter.ExpectedColumnDefinition;
+import org.apache.shardingsphere.sql.parser.integrate.jaxb.segment.impl.alter.ExpectedColumnPosition;
 import org.apache.shardingsphere.sql.parser.sql.segment.ddl.column.ColumnDefinitionSegment;
 import org.apache.shardingsphere.sql.parser.sql.segment.ddl.column.position.ColumnAfterPositionSegment;
 import org.apache.shardingsphere.sql.parser.sql.segment.ddl.column.position.ColumnPositionSegment;
