+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/jaxb/segment/impl/alter/ExpectedColumnPosition.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.integrate.jaxb.segment.impl.token;
+package org.apache.shardingsphere.sql.parser.integrate.jaxb.segment.impl.alter;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -27,15 +27,14 @@ import javax.xml.bind.annotation.XmlAttribute;
 @XmlAccessorType(XmlAccessType.FIELD)
 @Getter
 @Setter
-
-public class ExpectedColumnDefinition {
+public final class ExpectedColumnPosition {
     
-    @XmlAttribute
-    private String name;
+    @XmlAttribute(name = "start-index")
+    private int startIndex;
     
-    @XmlAttribute
-    private String type;
+    @XmlAttribute(name = "column-name")
+    private String columnName;
     
-    @XmlAttribute(name = "primary-key")
-    private boolean primaryKey;
+    @XmlAttribute(name = "after-column")
+    private String afterColumn;
 }
